dev: Eureka Server

+++ microservices/review-service/src/main/java/com/example/microservices/core/review/services/ReviewServiceImpl.java
@@ -2,11 +2,14 @@ package com.example.microservices.core.review.services;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.reactivestreams.Publisher;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.web.bind.annotation.RestController;
 import com.example.microservices.core.review.persistence.ReviewEntity;
 import com.example.microservices.core.review.persistence.ReviewRepository;
+import reactor.core.publisher.Flux;
+import reactor.core.scheduler.Scheduler;
 
 import com.example.api.core.review.Review;
 import com.example.api.core.review.ReviewService;
@@ -14,6 +17,9 @@ import com.example.util.exceptions.InvalidInputException;
 import com.example.util.http.ServiceUtil;
 
 import java.util.List;
+import java.util.function.Supplier;
+
+import static java.util.logging.Level.FINE;
 
 @RestController
 public class ReviewServiceImpl implements ReviewService {
@@ -21,13 +27,16 @@ public class ReviewServiceImpl implements ReviewService {
     private static final Logger LOG = LoggerFactory.getLogger(ReviewServiceImpl.class);
 
     private final ServiceUtil serviceUtil;
-    
+
+    private final Scheduler scheduler;
+
     private final ReviewRepository repository;
 
     private final ReviewMapper mapper;
 
     @Autowired
-    public ReviewServiceImpl(ReviewRepository repository, ReviewMapper mapper, ServiceUtil serviceUtil) {
+    public ReviewServiceImpl(Scheduler scheduler, ReviewRepository repository, ReviewMapper mapper, ServiceUtil serviceUtil) {
+        this.scheduler = scheduler;  
         this.repository = repository;
         this.mapper = mapper;
         this.serviceUtil = serviceUtil;
@@ -35,6 +44,9 @@ public class ReviewServiceImpl implements ReviewService {
     
     @Override
     public Review createReview(Review body) {
+
+    	if (body.getHotelId() < 1) throw new InvalidInputException("Invalid hotelId: " + body.getHotelId());
+
         try {
             ReviewEntity entity = mapper.apiToEntity(body);
             ReviewEntity newEntity = repository.save(entity);
@@ -48,10 +60,16 @@ public class ReviewServiceImpl implements ReviewService {
     }
     
     @Override
-    public List<Review> getReviews(int hotelId) {
+    public Flux<Review> getReviews(int hotelId) {
 
         if (hotelId < 1) throw new InvalidInputException("Invalid hotelId: " + hotelId);
 
+        LOG.info("Will get reviews for hotel with id={}", hotelId);
+
+        return asyncFlux(() -> Flux.fromIterable(getByHotelId(hotelId))).log(null, FINE);
+    }
+
+    protected List<Review> getByHotelId(int hotelId ) {
         List<ReviewEntity> entityList = repository.findByHotelId(hotelId);
         List<Review> list = mapper.entityListToApiList(entityList);
         list.forEach(e -> e.setServiceAddress(serviceUtil.getServiceAddress()));
@@ -63,7 +81,12 @@ public class ReviewServiceImpl implements ReviewService {
 
     @Override
     public void deleteReviews(int hotelId) {
+    	if (hotelId < 1) throw new InvalidInputException("Invalid hotelId: " + hotelId);
         LOG.debug("deleteReviews: tries to delete reviews for the hotel with hotelId: {}", hotelId);
         repository.deleteAll(repository.findByHotelId(hotelId));
+    }
+
+    private <T> Flux<T> asyncFlux(Supplier<Publisher<T>> publisherSupplier) {
+        return Flux.defer(publisherSupplier).subscribeOn(scheduler);
     }
 }
